refactor(ids): own shared link-health vocabulary below diagnostics
Move LinkTier and its order/wire pin unchanged; every caller imports ids directly so diagnostic records no longer depend on store.
